config: add cleanup_on_close config option to clean up fs when close snapshotter
Previously, the Close() interface of nydus snapshotter performed Cleanup()
on the filesystem. This behavior is not universal. In fact, on the
contrary, in most cases, the default behavior of Close() should not affect
the existing filesystem so that the filesystem can still work normally
after Reconnect(). While in some special scenarios, the filesystem needs to
be cleaned up with the close of the snapshotter.

So we add introduce a new config option called 'cleanup_on_close' to
control whether to clean up the filesystem under the nydus snapshotter when
the snapshotter is closed, and clean up the nydusd daemon and snapshot at
the same time.

The default value of 'cleanup_on_close' is false.

Signed-off-by: zhaoshang <zhaoshangsjtu@linux.alibaba.com>