converter: improve pack performance by fifo
In the previous convert implementation, the builder (nydus-image) would
dump the blob and bootstrap, and then the converter would pack the two
files into a tar format, which has a performance loss, and the patch
solves this problem by introducing fifo file.

The converter creates a fifo file, the builder will write the blob and
bootstrap data to the fifo as the writer side, then the converter read
from the fifo and write directly to content store for faster conversion.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>