log: fix logdir and output bug of snapshotter
There are three log related parameters, namely LogLevel, LogDir and
LogToStdout.

Previously, both LogDir and LogToStdout are directly passed to
nydusd daemon, while LogLevel is shared by snapshotter and
nydusd daemon, so actually we can't control the output direction and
file path of the snapshotter log. Now we fix it.

In order to prevent the already running business from being affected on
a large scale, we will not modify the log control method of nydusd
daemon here and continue to let nydusd share the configuration of log
with snapshotter.

Now the default path of nydusd log is (if DaemonMode is not none):
/PATH/TO/ROOTDIR/logs/NYDUSD_DAEMON_ID/stderr.log

And the default path of snapshotter log is:
/PATH/TO/ROOTDIR/logs/nydus-snapshotter.log

Signed-off-by: zhaoshang <zhaoshangsjtu@linux.alibaba.com>