filesystem: rename nydus to fs
Prepare for removing stargz filesystem implementation.

Since the stargz filesystem implementation is very similar
to the nydus filesystem implementation, much of the code
can be shared, for example the shared daemon related codes.

We are going to merge them together, this PR moves codes,
and breaks the current stargz implementation.

Signed-off-by: Yan Song <yansong.ys@antfin.com>