cache manager: fix cache manager failure in none daemon mode
Previously, both AddSnapshot and DelSnapshot will not be executed in
the none daemon mode. The fundamental reason is that the incoming
parameter is the imageID obtained from the daemon. Now we adjust the
function call path to bypass the daemon and get the imageID directly
from snapshot labels.

For AddSnapshot, we separate it from prepareRemoteSnapshot, run it
separately, and parse the imageID directly from the snapshot label.

For DelSnapshot, we advance it from Cleanup to Remove.

Signed-off-by: zhaoshang <zhaoshangsjtu@linux.alibaba.com>