fix: kill nydus-image process with timeout
In general, nydus-image does not run timeout, but in order to be
prepared, after all, there is such a risk in theory.
In addition, we found in some business scenarios that in extreme cases
(such as insufficient CPU resources and too large OCI layers), the
conversion time of nydus-image process may be calculated in hours.

Signed-off-by: zhaoshang <zhaoshangsjtu@linux.alibaba.com>