[Bugfix] Use the correct label to distinguish the lower layer from the upper layer
Previously, the unreliable (may not exist, such as when disable_snapshot_annotations = true)
label added by cri was used as the basis for judging the upper layer, then when the OCI image
without the label was started, it would mistakenly enter prepare for upper layer logic.

We noticed that "containerd.io/snapshot.ref" is a label added by containerd before calling
sn.Prepare(), and only exists in the lower layer snapshot, not in the upper layer, so we can
simply put the negative situation of lower layer as the upper layer.

Signed-off-by: zhaoshang <zhaoshangsjtu@linux.alibaba.com>