support reference type for nydus image
Add the new `enable_referrers` option in config.toml, to enables trying
to fetch the nydus image associated with the OCI image and run it.

This feature relies on the support of OCI reference type. Currently, only
Harbor (>=v2.8) registry and ORAS registry support this. Additionally, we
need to rely on containerd's referrer fetcher interface, related links:

https://github.com/containerd/containerd/pull/7644
https://oras.land/cli/6_reference_types/

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>