Manager: make nydusd daemon reach desired state
In the faileover process, if the new daemon is unexpectedly
killed before the new daemon is subscribed to by monitor,
the manager will not retry to start the daemon to ensure
that the new daemon runs successfully.

In this PR, the manager will retry to start the daemon
if the process is not running.

Signed-off-by: Bin Tang <tangbin.bin@bytedance.com>