release: Upload s390x tarball
This series is the very start to allow shipping tarball including
binaries for different architectures than x86_64.

The focus here is s390x, which is used by Confidential Containers, but
this can easily be expanded in the future.

Last but not least, for now we have *no* interest in publishing a
multi-architecture image.

Fixes: https://github.com/containerd/nydus-snapshotter/issues/548

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>