snapshot: get image reference from cri annotations for guest pull
The sandbox image (pause image) name cannot be stored in annotations by containerd,
so snapshotter is unable to retrieve it. This means that the source field of `KataVirtualVolume`
has to be set to "" to support guest pull in proxy mode now. However, once containerd fixes this bug
(https://github.com/containerd/containerd/pull/9419), I think the nydus snapshotter could be able to get the sandbox image name
from annotations directly. I recommend that we can support to obtain the image name through
"containerd.io/snapshot/cri.image-ref" in snapshotter. If this value is empty, the source should be set to "",
otherwise it should be set to the image name.

Signed-off-by: ChengyuZhu6 <chengyu.zhu@intel.com>