misc: Add basic function to detect the container runtime
Right now this is very basic and simply will let us know whether we're
running on a node that's using containerd or not.

This ended up with the move of parts of the logic on detecting
containerd and containerd configuration files being removed from a few
specific functions and being put directly into the main one.

In the coming patches of this series we'll expand it to also detect
other kubernetes flavours, such as `k3s` or `rke2`.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>