misc: Detect k0s and error out if used
Differently than `k3s` or `rke2`, getting `k0s` to work as expected is
quite tricky, as it relies on a drop-in configuration file, and I'm not
so confident we can easily drop-in a file with all the needed bits.

With this in mind, at least for now, let's just error out in case k0s is
used, as at least this will make sure that the users will at least have
a clue on what's been gone wrong.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>