go.mod: github.com/containerd/containerd v1.7.7, containerd/nri v0.4.0
Update to the first release of containerd to alias the log package to
the new containerd/log module.

This update of the containerd module also comes with an updated version
of github.com/containerd/nr, which has some breaking changes in the
interface definitions.

the optimizer-nri-plugin and prefetchfilees-nri-pugin pacakges had to be
updated due to [containerd/nri@9d86150] changing the interface to require
a context to be passed.

[containerd/nri@9d86150]: https://github.com/containerd/nri/commit/9d86150fce4318491481e354d0f34c6b76e8806e

Full diffs:

- https://github.com/containerd/containerd/v1.7.0...v1.7.7
- https://github.com/containerd/nri/compare/v0.3.0...v0.4.0

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>