converter: only enable tar-rafs conversion by env NYDUS_ENABLE_TAR2RAFS
In recently, we found the tar-rafs conversion lost the opaque file
handle: https://github.com/dragonflyoss/nydus/issues/1679

Compare to dir-rafs conversion, the tar-rafs conversion is still a
experimental feature, so we'd better disable it by default.

And provide an new env `NYDUS_ENABLE_TAR2RAFS` to allow users to enable it.

Signed-off-by: Yan Song <yansong.ys@antgroup.com>