[converter] Add MergeManifest option to tell converter when manifests are merged
When both the OCI and nydus manifests are merged with `nydusify --merge-platform` for instance, containerd will try to pull an unpack both images even if it doesn't have a nydus-snapshotter. Which results in the nydus image failing to be unpacked and a global image pull failure. One way to fix this issue is to use a config mediaType that containerd doesn't understand in the nydus image. This way it won't attempt to unpack it while the nydus snapshotter can still discover the image and pull it.

We use `application/vnd.nydus.image.config.v1+json` as the new config value only used if MergeManifest is set in the MergeOption.